backend/requirements.txt
Django==2.1.4
django-environ==0.4.5
djangorestframework==3.9.0
django-stdimage==4.0.1
pytz==2018.7
pycryptodome==3.6.6
django-sendgrid-v5
sendgrid==5.6.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
